@@ -1,21 +1,18 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net;
 using ServiceStack;
 
-namespace MyApp
+[assembly: HostingStartup(typeof(MyApp.ConfigureUi))]
+
+namespace MyApp;
+
+public class ConfigureUi : IHostingStartup
 {
-    public class ConfigureUi : IConfigureAppHost
-    {
-        public void Configure(IAppHost appHost)
-        {
+    public void Configure(IWebHostBuilder builder) => builder
+        .ConfigureAppHost(appHost => {
             appHost.CustomErrorHttpHandlers[HttpStatusCode.NotFound] = new SharpPageHandler("/notfound");
             appHost.CustomErrorHttpHandlers[HttpStatusCode.Forbidden] = new SharpPageHandler("/forbidden");
             
             Svg.Load(appHost.RootDirectory.GetDirectory("/assets/svg"));
             Svg.CssFillColor["svg-icons"] = "#2f495e";
-        }
-    }
+        });
 }
-
