@@ -1,18 +1,17 @@
-using System;
-using System.Collections.Generic;
-using Microsoft.Extensions.DependencyInjection;
 using ServiceStack;
 using ServiceStack.Web;
 using ServiceStack.Auth;
 using ServiceStack.Configuration;
 
+[assembly: HostingStartup(typeof(MyApp.ConfigureAuthRepository))]
+
 namespace MyApp
 {
-    // Custom UserAuth Data Model with extended Metadata properties
+    // Custom User Table with extended Metadata properties
     public class AppUser : UserAuth
     {
-        public string ProfileUrl { get; set; }
-        public string LastLoginIp { get; set; }
+        public string? ProfileUrl { get; set; }
+        public string? LastLoginIp { get; set; }
         public DateTime? LastLoginDate { get; set; }
     }
 
@@ -33,26 +32,17 @@ namespace MyApp
         }
     }
 
-    public class ConfigureAuthRepository : IConfigureAppHost, IConfigureServices, IPreInitPlugin
+    public class ConfigureAuthRepository : IHostingStartup
     {
-        public void Configure(IServiceCollection services)
-        {
-            services.AddSingleton<IAuthRepository>(c =>
-                new InMemoryAuthRepository<AppUser, UserAuthDetails>());
-        }
-
-        public void Configure(IAppHost appHost)
-        {
-            var authRepo = appHost.Resolve<IAuthRepository>();
-            authRepo.InitSchema();
-
-            //CreateUser(authRepo, "admin@email.com", "Admin User", "p@55wOrd", roles:new[]{ RoleNames.Admin });
-        }
-
-        public void BeforePluginsLoaded(IAppHost appHost)
-        {
-            appHost.AssertPlugin<AuthFeature>().AuthEvents.Add(new AppUserAuthEvents());
-        }
+        public void Configure(IWebHostBuilder builder) => builder
+            .ConfigureServices(services => services.AddSingleton<IAuthRepository>(c =>
+                new InMemoryAuthRepository<AppUser, UserAuthDetails>()))
+            .ConfigureAppHost(appHost => {
+                var authRepo = appHost.Resolve<IAuthRepository>();
+                authRepo.InitSchema();
+                // CreateUser(authRepo, "admin@email.com", "Admin User", "p@55wOrd", roles:new[]{ RoleNames.Admin });
+            }, afterConfigure: appHost => 
+                appHost.AssertPlugin<AuthFeature>().AuthEvents.Add(new AppUserAuthEvents()));
 
         // Add initial Users to the configured Auth Repository
         public void CreateUser(IAuthRepository authRepo, string email, string name, string password, string[] roles)
